[apiFirst/openAPi] Add & Modify & Changes different routes to the OpenAPI Architect
